Revert "use half() to convert float32 to float16, saving VRAM"
This reverts commit df5887e78e961a8fbd4b92fb7c84d0e06d2a1442.

+++ Script/train_ddd_pars_est_DiffPool.py
@@ -441,19 +441,17 @@ def main():
             )
 
         def forward(self, x, adj, mask=None):
-            s = self.gnn1_pool(x, adj, mask).half()
-            x = self.gnn1_embed(x, adj, mask).half()
+            s = self.gnn1_pool(x, adj, mask)
+            x = self.gnn1_embed(x, adj, mask)
 
             x, adj, l1, e1 = dense_diff_pool(x, adj, s, mask)
-            x = x.half()
 
-            s = self.gnn2_pool(x, adj).half()
-            x = self.gnn2_embed(x, adj).half()
+            s = self.gnn2_pool(x, adj)
+            x = self.gnn2_embed(x, adj)
 
             x, adj, l2, e2 = dense_diff_pool(x, adj, s)
-            x = x.half()
 
-            x = self.gnn3_embed(x, adj).half()
+            x = self.gnn3_embed(x, adj)
             x = x.flatten()
             x = self.linear_layers(x)
